Propose a new structure for the package
Historically, rally/plugins dir had bad structure. It was messy and hard
to understand which module relate to what rally component.

Since xrally_docker is a new package, which even was not released yet,
we can fix the issue and use the proper structure from the beggining.
It is easy to change and we do not need to do it in backward compatible
way.

Proposed structure:

  xrally_docker.common  -  a module for common things like utils,
                           service classes and etc
  xrally_docker.env     -  a module for Environment component.
  xrally_docker.task    -  a module for Task component including all
                           scenario, context, sla plugins
  xrally_docker.verify  -  a module for Veriication component